refactor: move equal and notEqual routes to validator.ts
